Movie Provider comlete, moved FetchMovieData to seperate class file ready for testing.
